Fix example for filtering in README.md
In the example for filtering, evt.target is null, so you cannot differentiate between the delete and edit buttons.  ctrl should instead be set to evt.to.  The example from the article (http://jsbin.com/yexine/6/edit?html,js) for filtering should also be edited. Currently, clicking the x brings up the edit dialog because only the default "else" behavior is being followed.