forceFallback changes
removed the default behaviour to append the cloned Element to the body.
now the cloned Element gets added in the same parent, with the addition
of the class defined in options.fallbackClass.
added the possibility to change the fallback class.
added the possibility to decide wheter the fallback should be cloned
into the same parent or to the document's body.