fix for sorting elements within a non-homogenus list of elements
For instance, say you are sorting tbody's within a table element.  There
can also exist tfoot and thead within the table element.  The _index()
function assumes that the list of elements within the parent is
homogenus (ie: just li's below an ol/ul).  This isnt the case when
sorting tbody under table.  This is edge case until you want to group
tr's under a table and sort on those groupings.