Allow event functions to be defined as functions AND strings.
The current behavior (in `master`) is to define the handling functions for events as strings. Default for `onSort` for example is `handleSort`. Commit 8cef762d11dbabb27e489727bcc8fd5dfe8d844d changed this behavior forcing handling functions to be defined as functions in the options. This breaks all existing code, resulting in a no-op even with the default options (which still define the default handlers as strings).

This commit restores previous behavior AND allows you to define your handlers as functions directly, if you’re so inclined.