Prevent ignored touches on Android Chrome.
Android Crhome will not trigger a click event if a user drags an item,
even if the item is dropped back into its original placement. This
means that whenever the "ignore next click" handling is activated, the
user's _next_ click (unrelated to dragging) will be ignored.

This commit solves this by never activating the "ignore next click"
handling on Android Chrome.